refactor: backend-structure
