+++ backend/schemas/stocks.py
@@ -1,0 +1,12 @@
+# backend/schemas/stocks.py
+from pydantic import BaseModel
+
+
+class StockCandle(BaseModel):
+    date: str
+    open: float
+    high: float
+    low: float
+    close: float
+    volume: int
+
